# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handler.Table:
    properties:
      id:
        type: integer
      qr_code_url:
        type: string
      restaurant_id:
        type: integer
      table_number:
        type: integer
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: API for Order System with user authentication and restaurant management
  title: Order System API
  version: "1.0"
paths:
...
  /api/restaurant/{restaurant_id}/table:
    get:
      description: Get all tables for a restaurant
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurant_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.Table'
            type: array
        "404":
          description: Restaurant not found
          schema:
            type: string
        "500":
          description: Error retrieving tables
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all tables
      tags:
      - Table
    post:
      consumes:
      - application/json
      description: Create a new table for a restaurant
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurant_id
        required: true
        type: string
      - description: Table data
        in: body
        name: table
        required: true
        schema:
          $ref: '#/definitions/handler.Table'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.Table'
        "400":
          description: Invalid input
          schema:
            type: string
        "404":
          description: Restaurant not found
          schema:
            type: string
        "500":
          description: Error creating table
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new table
      tags:
      - Table
  /api/restaurant/{restaurant_id}/table/{id}:
    delete:
      description: Delete a table
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurant_id
        required: true
        type: string
      - description: Table ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Restaurant or table not found
          schema:
            type: string
        "500":
          description: Error deleting table
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a table
      tags:
      - Table
    put:
      consumes:
      - application/json
      description: Update a table
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurant_id
        required: true
        type: string
      - description: Table ID
        in: path
        name: id
        required: true
        type: string
      - description: Table data
        in: body
        name: table
        required: true
        schema:
          $ref: '#/definitions/handler.Table'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Table'
        "400":
          description: Invalid input
          schema:
            type: string
        "404":
          description: Restaurant or table not found
          schema:
            type: string
        "500":
          description: Error updating table
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update a table
      tags:
      - Table
...
  /api/table:
    get:
      description: Get all tables for all restaurants belonging to the user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.Table'
            type: array
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Error retrieving tables
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all user tables
      tags:
      - Table
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"